# backend/src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "React Book 페이지를 위한 API입니다",
    "version": "1.0.0",
    "title": "BOOK API",
    "contact": {
      "email": "dmswlc19@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes": ["http"],
  "host": "localhost:5000",
  "basePath": "/api",
  "paths": {
...
    "/deepbrain/modelList": {
      "post": {
        "tags": ["DeepBrain"],
        "summary": "모델리스트",
        "description": "모델리스트 POST API",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/modelMessage"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "type": "object",
              "properties": {
                "statueMessage": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/deepbrain/makeVideo": {
      "post": {
        "tags": ["DeepBrain"],
        "summary": "비디오키",
        "description": "비디오키 POST API",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "language": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "clothes": {
                  "type": "number"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "object",
              "properties": {
                "statueMessage": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "type": "object",
              "properties": {
                "statueMessage": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/deepbrain/findProject": {
      "post": {
        "tags": ["DeepBrain"],
        "summary": "비디오 가져오기",
        "description": "비디오 가져오기 POST API",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "object",
              "properties": {
                "statueMessage": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "model": {
                          "type": "string"
                        },
                        "clothes": {
                          "type": "string"
                        },
                        "language": {
                          "type": "string"
                        },
                        "locale": {
                          "type": "string"
                        },
                        "process": {
                          "type": "string"
                        },
                        "video": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "type": "object",
              "properties": {
                "statueMessage": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "modelMessage": {
      "type": "object",
      "properties": {
        "succeed": { "type": "boolean" },
        "models": {
          "type": "array",
          "items": {
            "properties": {
              "id": {
                "type": "string"
              },
              "label": {
                "type": "object",
                "properties": {
                  "en": {
                    "type": "string"
                  },
                  "ko": {
                    "type": "string"
                  },
                  "zh": {
                    "type": "string"
                  }
                }
              },
              "expertise": {
                "type": "object",
                "properties": {
                  "en": {
                    "type": "string"
                  },
                  "ko": {
                    "type": "string"
                  },
                  "zh": {
                    "type": "string"
                  }
                }
              },
              "imgPath": {
                "type": "string"
              },
              "clothes": {
                "type": "array",
                "items": {
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "label": {
                      "type": "object",
                      "properties": {
                        "en": {
                          "type": "string"
                        },
                        "ko": {
                          "type": "string"
                        },
                        "zh": {
                          "type": "string"
                        }
                      }
                    },
                    "imgPath": {
                      "type": "string"
                    },
                    "intro": {
                      "type": "string"
                    }
                  }
                }
              },
              "language": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "gpuType": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}